[更新]整理setting page 程式碼
