+++ src/routes/setting/setting.style.ts
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  padding: 4rem 4rem 0;
+  color: white;
+`;
+
+export const Form = styled.form`
+  display: flex;
+  flex-direction: column;
+  background-color: #18181b;
+  padding: 2rem 2rem 0;
+  border-radius: 0.4rem;
+`;
+
+export const Title = styled.h2`
+  font-size: 2.5rem;
+  margin-bottom: 1rem;
+  font-weight: bold;
+`;
